tokenserver: Add identityset.Set struct that will be used by delegation rules.
The delegation rules are formulated in the language of sets ("these rules apply
to the query iff such and such sets are subsets of the sets specified in the
rule").

They are also defined as a mix of IDs, groups and special '*' character that
means "any identity".

This struct implements the corresponding set structure and all operations that
will be needed later.

R=nodir@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2420493002