Fix sparse index handling, add index params.
Fix a bug in index archival where sparse indexes without a last log
entry will act as if they have no logs.

Augment the index protobuf with stream-wide parameters, namely the last
prefix and stream indexes in the stream and the total number of entries
in the stream. This will help supplement index information for sparse
indexes. This is also backwards-compatible with indexes without these
parameters.

Finally, add a tool to help query and interact with archived data
records. This was useful in diagnosing this bug, fixing this bug, and
is expected to be useful for future purposes.

BUG=chromium:654880
TEST=local,production
  - Tested against production data. Tool without patch exhibits
    production bug. Tool with patch successfully loads the stream.

Review-Url: https://codereview.chromium.org/2422393002