LogDog: Fix archival Get/Tail implementations.
The sparse index tail algorithm assumed that indexes always had an entry
for the last log in the index. This is not necessarily true for spare
indexes. Conseqently, archive Tail could return a non-tail entry in that
situation.

This also fixes the Get and Tail to be more robust against malformed index
and stream data files, to the point where it will happily operate
without an index if needed now.

However, this is actually a good idea, so update index generation to
always include a final entry to the index. Tail optimizes for this case.

This introduced yet another situation where the storage layer was
unmarshalling LogEntry protobufs, then passing the binary data up to the
higher level to unmarshal again! To prevent this from introducing
performance regressions, we now have the storage layer kick up a lazy
unmarshal-at-most-once struct instead of raw data. This will let higher
levels take advantage of storage layers that have to unmarshal
protobufs.

Finally, log archival fetching had no supporting unit tests. Write some,
with very high coverage.

TBR=nodir@chromium.org
BUG=chromium:654880
TEST=unit

Review-Url: https://codereview.chromium.org/2435883002