LogDog: Update client/bootstrap to generate URLs.
Add the ability for a populated Bootstrap to generate viewer URLs.

Update streamclient.Stream to be able to emit its stream properties. This is
useful in general, notably by the Bootstrap's URL generation functions.

This change involves modifying streamproto.Properties to remove the false
assumption that they are safe to copy/modify. To this end, places where
this assumption was made now properly Clone() the Properties to take
possession of them, and Properties are now passed around as a pointer to
make this clear.

BUG=chromium:659291
TEST=unit

Review-Url: https://codereview.chromium.org/2456953003