Reduce overhead of getting one single number from mathrand.
If the context doesn't have math.Rand installed (which is the case in all
non-testing code), Get(...) generates new rand.Rand object.

This is very expensive if its going to be used only to get one sample (up to
400x more expensive than just using math/rand global RNG directly).

Provide convenience functions that handle empty context in a more efficient
way.

R=nodir@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2466923003