server/templates: Use function for DebugMode.
Currently the DebugMode option is a boolean set in advance.
This precludes it from consulting Context services to determine if it
should be run in debug mode.

Replace the boolean with a Context-aware function. This function happens
to match the signature of info.IsDevAppServer, which is exclusively how
this option is used at the moment, making it easy to plug in.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2466853003