Update environ.Env to hide its implementation.
Currently, environ.Env exposes itself as a map with methods. However,
directly manipulating the map is discouraged, since it exposes a
counter-intuitive implementation detail to the user.

Replace this with a struct containing the map. This hides the
implementation detail and forces the user to go through the standard
function set to manipulate the environment.

Add a Make function so users can create an empty environment from a map.

Add a Map function to return the more-intuitive key/value map.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2467293003