Add callbacks to query batch filter.
Currently, the query batch filter transparently batches queries.
However, users wanting to act on query data as it comes in have no safe
point of intervention: if they operate on the data inside their
callback, they may be operating in the middle of an ongoing query,
causing it to time out. If they operate afterwards, they will have to
buffer the full set of query results.

This adds optional callback(s) to the batch query filter. These
callbacks will be invoked in between batches.

BUG=None
TEST=None
R=iannucci@chromium.org

Review-Url: https://codereview.chromium.org/2484513002