cipd: Add '-compression-level' option.
The default level is still 5, as it was. It will likely be changed to 1 in the next CL.

Fastest compression level is good enough for us, since we almost never care
about storage, but we do care about the latency of jobs that install stuff.

Comparison of infra_python.cipd sizes per level:
0 - 75 Mb (uncompressed)
1 - 27 Mb (the proposed new default)
5 - 23 Mb (the old default)
9 - 23 Mb (paranoid value)

Installation times for level >=1 aren't much different (~1.5 sec) on my machine.

Stats for 'skia/bots/skimage:version:12' package (421 Mb of various images):
0 - 421 Mb, 1.4 sec
1 - 391 Mb, 3.1 sec
5 - 385 Mb, 7.9 sec

All tests are done on MacBook Pro (Mid 2014), so they are not very
representative :)

R=iannucci@chromium.org
CC=kjlubick@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2476253004