Make mathrand consistently concurrency-safe.
A *math.Rand object is not concurrency-safe. It must be locked around in
order to safely use.

The "mathrand" abstraction layer, however, does not properly implement
this. It has:
- Context method which:
  - Returns a new instance if a *math.Rand is not explicitly set in the
    Context (concurrency-safe on virtue of new allocation).
  - Can return the Context-global instance (not safe) if one is
    installed.
- Global functions which:
  - Use math/rand's safe global methods if a *math.Rand is not installed in
    the Context.
  - Can return the Context-global instance (not safe) if one is
    installed.

We fix this by introducing a generic "Rand" interface and exposing that
to the user. Internally, the Context-global Rand instance is protected
behind a mutex.

Fixes #35.

BUG=None
TEST=None
R=iannucci@chromium.org, vadimsh@chromium.org

Review URL: https://codereview.chromium.org/2492223003 .