impl/cloud: Add support for "memcached" memcache.
Add "memcache" service support to "impl/cloud" backed by a "memcached"
cluster.

This enables a full dscache-backed datastore to be implemneted using
"impl/cloud", as well as other services that need memcache to operate.

Update the "impl/cloud" interface to start with a service Config and
install services from that.

BUG=None
TEST=local
  - Started "memcached" server, tested against it.

Review-Url: https://codereview.chromium.org/2513253002