client/isolate: add TarArchiver
The tar archiver combines many small files into fewer large tar files,
keeping track of the digest as it builds each tar.

Each created archive is held in memory (and not committed to disk). The
tar archiver maintains a cap on the number of resources used by only
allow a set number of tar archives to be created at a single time.

Issue luci/luci-go#9

BUG=59899

Review-Url: https://codereview.chromium.org/2536063002