LogDog: Add signed GS URL fetching.
Allow a "Get" RPC request to additionally request a signed URL to the
archive log entry stream, if available.

To really support this, we restructure (for the better) the internal
service API and the Coordinator test framework to enable fetching bound
storage instances instead of the motley assortment of composite services
that was previously implemented.

BUG=chromium:669994
TEST=unit

Review-Url: https://codereview.chromium.org/2538203002