Add GAE datastore caching layer implementation.
Add an AppEngine datastore-backed caching layer. This layer allows
arbitrary packages to register a cache handler and use a
datastore-backed cache for its data. It is suppored by a cron task which
refreshes and/or prunes cache entries periodically.

The initial implementation's schema is shard-able, but its initial user
(config service) is not expected to need shards, so full shard support
is not implemented. This implementation is forward-compatible with such
support, though, should it be added in the future.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2483233003