Rewrite LogDog log viewer app.
Rewrite the LogDog log viewer app in TypeScript, and integrate it into
the Gulp viewer. This adds many features, including:
- Actual TypeScript 2.0 versions of the LogDog libraries.
- The ability to perform log tails and manual (on-demand) log loading.
- Much faster log rendering (by default).
- Prefetch initial log chunk so it feels more responsive.
- Better model/view/controller separation.
- Transient HTTP and RPC retries.
- A tighter authentication loop, hopefully better now.
- Better support for mobile apps.

The current state of things is hybrid JavaScript / Polymer / TypeScript
because too much change in one CL is bad. The minimum bridges were built
between the JavaScript libraries that weren't rewritten and the
TypeScript code that uses them.

The TypeScript code is smaller, cleaner, and (obivously) stronger typed.
This should lead to its significantly-increased enhancement and
maintainability.

Polymer integrates TypeScript by including it during vulcanization. This
requires the Gulp build system to include an intermediate "build"
directory step, where the original sources are copied over and the
transpiled JavaScript code is generated.

BUG=chromium:667990
TEST=local
  - Ran the app locally, seems to work under all expected circumstances.

Review-Url: https://codereview.chromium.org/2543323004