LogDog/Butler: Explicitly pass Coordinator host.
Currently, the Butler infers the Coordinator host from the configured
LogDog output. This changes it to be explicitly passed via command-line.
This has two practical impacts:

Firstly, it avoids a problem where, if the Output is pointing to a
service-specific backend (e.g., services-dot-<host>), we are forwarding
that backend to subprocesses, which are then using it to construct their
viewer URLs. This lets us explicitly forward the host value that the
subcommands use.

Secondly, the coordinator host code path is currently only available
when using the "logdog" output. Making it an explicit parameter will
make it accessible to other outputs for debug or simulation testing.

BUG=None
TEST=None
R=iannucci@chromium.org

Review-Url: https://codereview.chromium.org/2575113002