client/isolate: only use the estimated tar size for splitting
It turns out that the estimated tar size is occasionally incorrect for
reasons that elude me, which causes swarming to fail ^.
So, the size calculation should only be used as the basis for splitting.

Instead, measure the size of the tar while calculating its SHA1 hash,
and use that when constructing the isolated JSON and communicating with
the isolate service.

^ eg. "Failed to fetch
a9dcea82cb5702455c216dc3be623c19764605ab: Incorrect file size: expected
9999872, got 10002944"

R=mcgreevy@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2576313002