Create unbuffered Tumble entry point for LogDog.
Add RegisterInTransaction, an entry point for Tumble that initiates a
standard datastore transaction, performs an operation, and registers the
ensuing mutation against it. This is different from RunMutation in that
it doesn't require a Mutation as a starting point and also doesn't
create a transaction buffer.

Leverage this in LogDog's "registerStream" to reduce the overhead of
stream registration.

BUG=chromium:675485
TEST=unit

Review-Url: https://codereview.chromium.org/2592753002