Fix BigTable multi-entry row cached Tail.
When performing a Tail, LogDog has to scan through all of a log's
BigTable rows to identify the last contiguous entry. In order to
optimize this, the BigTable storage implementation caches the last
contiguous tail index in memcache.

BigTable rows are keyed on their LAST log entry index, not their first,
so the cached log index was the last index in the row. If the row
contained multiple log entries, this violated an earlier continuity check,
which was expecting the first index in the row, not the last, to match the
cached index.

We fix this by caching the first index in the row, matching continuity
expectaitons.

BUG=chromium:678831
TEST=unit,local
  - Added regression unit test, passes now.
  - Use multi-round Tail on `logdog_bigtable_test` before and after,
    verify that it works now.

R=vadimsh@chromium.org

Review-Url: https://codereview.chromium.org/2611253005