Fix improper use of Stop in datastore.
The datastore.Stop sentinel error should be used by callbacks to
indicate an end of iteration. We were improperly using it
impl/memory's query handler in response to reaching a limit (appropriate
thing here is to just naturally stop the query) and datastore.Batcher's Run
handler as an unnecessary and erroneous safeguard.

BUG=chromium:679064
TEST=local
  - Ran unit tests here and in luci-go, work now.

Review-Url: https://codereview.chromium.org/2616243002