Fix race in treapstore test.
The treapstore test currently asserts that at least one read happens.
The first round, however, it is possible that no writes happen before
the read round finishes. Adjust the test to accommodate this.

The check itself is not very useful. The purpose of this test is to
expose a bunch of potential races and assert that they aren't actually
races.

TBR=vadimsh@chromium.org
BUG=chromium:675485
TEST=local
  - for i in `seq 100`; do go test -cpu=1,2,4 ./common/data/treapstore;
    done

Review-Url: https://codereview.chromium.org/2613153007