Implement a server-side config service interface.
This implements "server/config", a service-specific config service
interface. This interface is noticably different from the
"common/config.Interface" in the following ways:
- It accepts an authority: on whose behalf is the config request being
  issued?
- Defines a greatly-reduced interface, allowing for much simpler cache
  implementations.
- Allows for format translation, enabling implicit conversion and cache
  storage of configuration data in a different format from its native
  config service representation. Specifically, this targets storing text
  protobufs as binary.

Ultimately, we will delete the equivalent packages out of
"common/config", leaving it as a client endpoint to the config service.
Updates to services, notably "appengine/gaeconfig" will follow.

BUG=chromium:674378
TEST=unit

Review-Url: https://codereview.chromium.org/2580713002