Tumble operates on per-namespace shards.
Review Tumble's namespace model. Currently, a given Tumble shard will
perform batches of parallel operations, one per namespace, in response
to a given task. If there is no work to be done, the shard will
terminate immediately.

This patch revises this. Now, any given shard-processing HTTP
handler processes a single namespace. This is determined by the
namespace to which the task queue task that started the handler is
bound. If there are multiple namespaces that need a shard processed,
they will kick off separate tasks in separate namespaces and be
processed by separate handlers. The result here is that individual
namespace operations are now completely independent.

Additionally, make it so that a given shard will not terminate
immediately if there is no work. Instead, define a tunable parameter,
MaxNoWorkDelay. If there is no work, the shard will sleep for some
period of time, exponentially backing off up to MaxNoWorkDelay. This
will ensure that a given shard will not repeatedly respawn only to do no
work.

We still keep the polling cron. When that kicks off, any available shard
data is processed, but the shard will not loop if there is no work.

The net expected effect of this is:
- More individual handlers will run.
- Handlers will only run for active shards in active namespaces.
- Mutations within a namespace are now competely independent from other
  namespaces (with the exception of global quotas).
- Idle namespaces (namespaces with few/no mutations) will waste few
  cycles, as they either will not be scheduled, or will be caught in an
  exponetial backoff delay most of the time.

Existing tasks bound to the special task namespace will be disregarded.
This should be fine, since the only system that is leveraging them is
LogDog, and it has enough volume to cause its shards to be processed
regardless.

BUG=chomium:678896
TEST=unit

Review-Url: https://codereview.chromium.org/2628683002