tumble: If there is work always loop.
Fix a case where when a Tumble handler is configured not to loop
(launched by cron) but there still may be work to be done, it would stop
after one iteration. Now, in this case, it will continue to loop until
there is no work to be done, but won't loop past that polling for more
work.

BUG=chromium:678896
TEST=None

Review-Url: https://codereview.chromium.org/2627783004