Integrate TypeScript into luci/web.
Augment the base luci-go/web Gulp scripts and layout to include
TypeScript configuration parameters and compiling integration.

This change replaces "build.py" with a subcommand-aware "web.py", which
acts as an entry point to the web building framework. Notably, this adds
the "web.py presubmit" subcommand, invoked by PRESUBMIT, which will
perform linting as part of the repository's presubmit process.

BUG=chromium:678673
TEST=None

Review-Url: https://codereview.chromium.org/2611153003