datastorecache: Add Locker field, public settings.
Add the Locker field to datastorecache. Previously, the Locker was not
configurable, and defaulted to memlock. However, locking is optional for
fetching configuration data. This decouples datastorecache from the
requirement that it has a working memcache, allowing it to be used by
other datastore implementations such as cloud.

This also changes the names of the gaeconfig datastore cache settings to
be public so that they can be used by external packages that want to
process the gaeconfig settings.

BUG=chromium:682465
TEST=unit

Review-Url: https://codereview.chromium.org/2645763002