Refactor homedir use for auth secrets directory.
The homedir package is a third party dependency which provides access to
a user home directory without a dependency on cgo. The home directory is
used by the auth package as a location to create a cache for tokens.

This change refactors the secret directory to be associated with the
authcli package instead of the top-level common/auth package.

R=maruel@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2644403002