Refactor how oauth token cache is organized.
Also add more docs and tests.

The previous implementation had a number of shortcomings:
  * It used not-very-descriptive digest-based key for the cache file name, and
    didn't write any details about the origin of the token into the cache file.
    As a result, the cache directory ends up looking like a mess with a bunch of
    randomly-looking files.
  * It was possible to have a collision in names of the cache entries.
  * There was no automatic cleanup of old entries.
  * Cache key for service account JSONs was partially derived from the location
    of the key file, not its content (key ID). It was a hack.

New implementation stores all cached tokens in a single nice and tidy JSON file.

This allows to use structured cache keys, so we no longer lose information about
cache entry origins. It is also makes easier to remove old entries from
the cache.

The downside is a possibility of some token update being lost if the cache file
is being updated concurrently. This is bad only if the lost update is a cache
entry with new refresh token. This can happen only if "login" operation races
with some other operation that updates the cache. "login" happens very-very
infrequently, and only on end-user machines.

On bots (that exclusively use service accounts), a lost write to the cache is
merely an inefficiency: next time the process starts, it will have to redo
OAuth2 flow (1 HTTP request) to get a new access token in place of the lost one.

Once this is deployed, it will effectively "logout" everyone. At some point
later (when we are sure no code that uses old cache structure is running), we'll
add a change that cleans up old *.tok cache items.

R=iannucci@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2659263002