[cipd] Add support for windows file attributes.
This allows us to successfully package and restore the 'hidden' and
'system' file attributes for windows files. This is particularly
important becuse msys/mingw uses the 'system' flag to indicate that
certain files should be scanned e.g. for symlink processing, which means
that we can now package git-for-windows and deploy it with cipd.

R=vadimsh@chromium.org,dnj@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2696443004