Fix LogDog client queries without state.
LogDog is allowed to issue queries without retrieving state. However,
the client library considered a lack of state an error. Update it to
permit this.

Update Milo's tests to properly reflect the missing state for its
queries so the client library would have properly failed before, and now
properly succeeds.

BUG=chromium:687625
TEST=unit
R=hinoka@chromium.org

Review-Url: https://codereview.chromium.org/2697223002