milo: Enable Swarming LogDog log loading.
Enable Swarming builds to load directly from Milo annotation streams
when the build contains a LogDog "log_location" tag.

To do this, Swarming parses the tags when it fetches a task result. If
the tags include a "log_location" tag, and that location is LogDog, it will
terminate any log fetching (if applicable) and, rather than parse
Swarming logs through Annotee, load the annotation protobuf instead.

To enable this, we introduce the concept of a LogDog URL:
logdog://host/project/prefix/+/name

This URL is handled in the LogDog "types" package.

BUG=chromium:692220
TEST=unit

Review-Url: https://codereview.chromium.org/2695383002