Kill the correct session when relogining.
If OpenID callback was executed with already active session, it closes it and
opens a new one. Unfortunately, there was a typo-like bug there and it was
closing the new session instead (while still putting its cookie). It resulted
in a session cookie that points to closed session (this is treated as not logged
in).

Also add some logging for debugging and simplify enabling OpenID authentication
on devserver (useful for debugging).

R=nodir@chromium.org, hinoka@chromium.org
BUG=689270

Review-Url: https://codereview.chromium.org/2697923006