auth: Slightly change semantics of InteractiveLogin to make more sense.
It used to always trigger an interactive login in Transport(...) and other
factories, even if there is a cached refresh token already.

Now it merely indicates that it is OK to interact with the user to get a token,
but it won't always trigger a login (there separate Login() call for that, if
needed).

Also fix some nits in authcli.go.

R=iannucci@chromium.org, nodir@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2701293002