Butler: Infer host from "coordinator-host" flag.
Currently, the LogDog Coordinator host can be explicitly specified as a
top-level flag. This is redundant with the host, which is also specified
in the LogDog "output" module.

The reason for this is that some configurations may wish to direct the
Butler to a services module within the host (e.g., service-dot-example.com).
However, the host must still be available for viewer link generation.

Rather than specify it multiple times, we will use only the top-level
host flag. The production output, "logdog", can take an additional optional
service name which, if provided, can be used to construct the
"service-dot-host" host for service engagement.

BUG=chromium:694911
TEST=None

Review-Url: https://codereview.chromium.org/2715443002