Milo: Handle missing / transient LogDog failures.
If a build is labelled as having a LogDog stream, Milo will
unconditionally try and load that stream. This falls apart if the build
has either not started yet, or if the LogDog annotation stream is not
available due to natural ingest delay.

This will prevent stream loading if the build hasn't started yet, and
will retry LogDog transient and not-found errors prior to failing
absolutely.

BUG=chromium:692245, chromium:694913
TEST=None
R=hinoka@chromium.org, nodir@chromium.org

Review-Url: https://codereview.chromium.org/2717623002