Move GCS handling for isolate fetch into argument.
When the isolatedclient performs a Fetch operation, it first queries
isolate. When the blob is small, isolate returns the contents directly.
When the blob is large, however, isolate returns a URL for accessing the
blob from GCS, and then immediately fetches the blob.

This moves the GCS fetch into a function which can be defined when the
isolatedclient is created. This facilitates other methods of access.

R=maruel@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2719613002