Milo: In LogDog-only mode, only render LogDog link
Mirror recent BuildBot template changes: in LogDog-only mode, only
render the LogDog links, since the BuildBot links are useless.

To do this, we remove the explicit idea of links having alias sets.
Instead, links are passed around as LinkSet, which are a set of links
that should appear on the same row as each other. Any given link in the
set can be marked as an alias, meaning that it should show up
[like this].

Relaxing this structure lets us compose link sets based on whether or
not we're in LogDog-only mode.

BUG=chromium:698375
TEST=expectations

Review-Url: https://codereview.chromium.org/2733633003