Butler stream servers can generate client address.
Allow a StreamServer instance to generate client addresses. These
addresses can be processed by the stream client package to generate a
client for the corresponding StreamServer.

This also adds tests for UNIX and Windows stream server implementations.
These tests have the bonus of actually tying together this Address()
string to the "streamclient" package, confirming that they are
functional together.

In order for this connection to be made, the circular dependency that
"streamclient" had on "/logdog/client/butler" had to be broken. This is
done by moving the "local" client implementation into the "butler"
package space. This makes sense, since this implementation is an
internalization of the stream server.

BUG=chromium:698768
TEST=None

Review-Url: https://codereview.chromium.org/2737603003