Add TCP4/6 stream server support to LogDog Butler
Add the ability for the LogDog Butler to listen on TCP4 and TCP6 ports.
Those ports will, by default, bind to the localhost address.

This enables another method of Butler connectivity, and opens up the
possibility for a single Butler host to receive connections from other
systems (e.g., "adb", Docker, etc.).

Update the Butler's client to be able to connect to TCP4 and TCP6 ports.

BUG=None
TEST=local
  - Tested "tcp4" and "tcp6" stream server modes.

Review-Url: https://codereview.chromium.org/2724693003