Fix shadowing bug.
The local err declared in the useDB if branch shadowed the named err
return value.  This caused the naked return at the end of the function
to always return nil (in the useDB case) even if there was an error
during upload.

This change also reworks some of the test code to test that we actually
get errors when all retries fail. The previous tests checked that we
succeed eventually in the face of flakiness, but did not verify that
this success was due to retries, and not simply erronous reporting of
success.

Review-Url: https://codereview.chromium.org/2731933002