Add LUCI operation type, integrate cancellation.
Add the LUCI Operation type, which represents a single operation and can
be cancelled. The operation provides utility methods to cancel Promises
and to allow "on cancel" callbacks to be registered, enabling a generic
cancellation interface.

This will provide an easy interface for high-level code (e.g.,
the LogDog stream viewer) to cancel asynchronous operations efficiently.
It will also allow intermediate stacks to respond to or ignore
cancellation.

Integrate Operations into RPC client, retry, and sleep promise types so
that they can respond to an Operation's cancellation.

BUG=chromium:678673
TEST=None

Review-Url: https://codereview.chromium.org/2738823006