vpython: Add VirtualEnv creation package.
Add the "venv" package, which is responsible for creating and managing
collections of specified VirtualEnv environments.

This includes the ability to resolve, create, load, and manage (pruning)
an enviornment.

This package includes unit tests which assert the correctness of the
environment. These, in turn, leverage a few binaries (sadface). However,
the binaries are small, and this seems like a reasonable thing to do
given the subject under test.

BUG=chromium:691818
TEST=unit

Review-Url: https://codereview.chromium.org/2699063004