Fix GAE deployment artifact pathing.
luci_deploy was recently updated to use "gcloud" for artifact
deployment. However, this was not tested, and had a bug where the base
name of the artifact, rather than the full path, was passed to the tool.

Also, fix the untested "deploy_appengine" GAE management subcommand so that
it actually works.

TBR=iannucci@chromium.org
BUG=None
TEST=local
  - Ran "deploy_appengine" and observed successful deployment of a GAE
    configuration.

Review-Url: https://codereview.chromium.org/2745203002