Switch from "winio" to "npipe".
This is a hopeful fix to the linked bug. The problem is Go runtime panic
that manifested in coordination with an upgrade to Go 1.8. There are
several GitHub issue threads referencing similar issues with Windows and
"winio".

We have to disable the test because "npipe" is actually not a fully-
correct Go Listener implementation, and the test fails because of this.
This is okay, though, as this code has run for quite a while on Windows
using "npipe" without issue.

BUG=chromium:702105
TEST=None
R=vadimsh@chromium.org

Review-Url: https://codereview.chromium.org/2770013003