Promote LogDog aliases for BuildBot JSON data.
Generalize the LogDog alias-to-link promotion logic and apply it to the
BuildBot JSON endpoints. This will cause LogDog-only builds to omit
junk BuildBot log links in favor of the LogDog link aliases, bringing
the JSON in line with the rendered build.

Because Gatekeeper uses Milo for BuildBot step information, this will
incidentally cause Gatekeeper to include LogDog links in its failure
e-mails instead of the junk BuildBot links.

BUG=chromium:701201
TEST=expectations
  - Expectations, which already expressed this logic, didn't change.

R=hinoka@chromium.org

Review-Url: https://codereview.chromium.org/2774343003