vpython: Prefix PATH with Python bin directory.
Have "vpython" prefix the PATH environment variable with the "bin"
directory of its VirtualEnv. This is equivalent to what an activation
script would do.

This makes it so that scripts run by "vpython" that run "python"
automatically run the VirtualEnv's "python", and scripts that depend on
tools packaged in their wheels (e.g., "coverage") will have those tools
available on PATH.

BUG=None
TEST=local
  - Ran "vpython", confirmed PATH contains the VirtualEnv "bin" directory.

Review-Url: https://codereview.chromium.org/2803853003