token-server: Add a mechanism to prevent potential tokens misuse.
We use protocol buffers for token serialization, and we protect token integrity
via RSA-SHA256 signatures produced via SignBytes GAE API. That has following
consequences:
  1. We use same private key for signing all tokens, regardless of their kind.
  2. Token bodies are actually very similar structurally, since they are just
     proto messages that use smallish tag numbers (1, 2, 3, ...) as dictionary
     keys.
  3. It is conceivable that a token of kind A can be recognized as valid token
     of kind B with "wrong" data: they are both signed by the same private key,
     and they are both similar looking proto.

To avoid this situation we introduce a notion of "signing context" string. It is
a per token kind string that is prepended to the token blob before it is signed
or its signature is checked. We essentially seed SHA256 differently for
different token kinds.

This mechanism is not enabled yet for existing token kinds, because it will
break signatures of the in-flight tokens.

R=iannucci@chromium.org, nodir@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2789303002