Move GCS handling for isolate push into argument.
When the isolatedclient performs a Contains (preupload) operation, the
server returns the missing items from the store. Larger items are then
pushed to a GCS signed URL, generated by Isolate.

This moves the GCS push into a function which can be defined when the
isolatedclient is created. This facilitates other methods of access.

R=maruel@chromium.org, vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2802733005