vpython: spec v1, PEP425 tag resolution.
"vpython" now derives the PEP425 tags for the current Python
environment, and can use them in CIPD package specifications. This is
done by creating a vanilla (no additional wheels beyond VirtualEnv)
enviornment and using it to probe the PEP425 tags.

Tags are retained in the environment, and can be loaded; therefore, the
probe only needs to happen once per environment.

This includes some additional updates:
- Update "vpython" enviornment:
  - Environment specification and completion flag now incorporate
    "vpython" API version.
  - Updating version invalidates all environments.
  - Instead of writing enviornment specification to the environment, an
    environment description protobuf (which includes the specification) is
    written.
  - Environment now includes Python PEP425 tags associated with the
    environment.
- Fix filesystem.Touch to not truncate files.

BUG=None
TEST=local

Review-Url: https://codereview.chromium.org/2813883006