Cleanup PEP425 implementation.
The PEP425 implementation internalizes a lot of the details of how
PEP425 tags get applied. This CL simplifies this a bit.

Primarily, the full VirtualEnv runtime environment is now exported as a
protobuf to the package loader. This includes the tags. It is up to the
package loader to determine how this environment affects loading
packages.

The CIPD package loader offers a callback that allows its owner
to augment its template set based on the enviornment. This will let the
actual entry point specify how packages will be named, and how package
name resolution can be affected by the runtime environment. This makes
sense, since it brings the package resolution close to the party
responsible for creating the packages in the first place.

BUG=None
TEST=local

Review-Url: https://codereview.chromium.org/2817163002