Enable Python spec verification.
Add a "verify" subcommand that verifies the integrity of a specification
file and the ability to resolve packages against the configured
PackageLoader for a set of preconfigured verification template
parameters.

This can be used in PRESUBMIT and while developing specification files
to ensure that the specification will run on all targeted systems.

BUG=None
TEST=local

Review-Url: https://codereview.chromium.org/2815353002