auth: Use signJwt methods instead of signBlob for "act as" feature.
It makes the protocol slightly faster.

The problem with signBlob is that it can't be used to sign payloads that contain
the signing key ID inside them, because this key is not known prior calling
singBlob. The key is picked by signBlob internally.

JSON Web Tokens support optional field in the header that specifies the signing
key ID. Presence of this field makes tokens slightly faster to validate, because
a verifier doesn't have to enumerate all possible signing keys anymore.

JSON Web Tokens headers are signed along with the body, and thus we can't insert
signing key ID in the header and use signBytes to produces signed token.

signJwt RPC solves this problem. It is a specialization of signBlob RPC,
specifically for signing JSON Web Tokens. In understands the structure of JWTs
and knows how to produce correct header (that includes the signing key ID), sign
"<header>.<claimset>" and return the resulting JWT.

Also, since ClaimSet is now part of gcloud/iam API, copy it outside of
golang.org/x/oauth2/jws, since this package is marked as "not intended for
public us".

R=iannucci@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2819283002