environ: Fix key case, add GetEmpty and Remove.
The "environ" package does not ignore the case of enviornment variable
keys on Windows. This patch fixes this bug by allowing an Env to be
configured to be case insensitive, and having Windows systems default to
"true".

Creation methods "System" and "New" will both default to the case
sensitivity of the current operating system.

Additionally, this patch adds Remove, which allows for the deletion of
enviornment variables, and GetEmpty, which offers an interface similar
to "os.Getenv".

BUG=chromium:714165
TEST=unit

Review-Url: https://codereview.chromium.org/2833073004