environ: Disallow empty keys, values must have "="
Fix a bug in common/system/environ where environment variables without
values were exported without an "=". This is apparnetly fine on POSIX
platforms, but makes Windows fatally unhappy.

Also enforce that a key must have at least one character in it by
dropping zero-length keys, which are invalid on all platforms. This is
not attempting to perform a sanity check on the keys; rather, it is
avoiding a specific known failure case.

BUG=None
TEST=unit

Review-Url: https://codereview.chromium.org/2838743004