vpython: Fix race, more aggressive lock.
Fix a race in vpython where processes initialized while another process
is setting up the initial VirtualEnv would not load that VirtualEnv's
environment stamp protobuf.

Normally, if a completion flag exists, the process will load the
environment stamp protobuf for that environment. However, if the
completion flag doesn't exist, "vpython" will try and obtain a lock and
then, once obtained, create the enviornment. If the completion flag
exists when the lock is obtained, it means the environment was created
by another process while waiting for the lock, so we accept that
environment. However, in this case, we do NOT load the environment stamp
file. Fix this so that the file is actually loaded.

Additionally, decrease the lock polling interval from 5 seconds to 10
milliseconds, It's a really cheap operation (syscall) and there's no
reason not to be aggressive about it.

Additionally, in between waiting for the lock, check for the completion
timestamp. If it is created before we actually obtain the lock, then we
can skip waiting for the lock altogether.

Finally, increase default logging level to error. Some processes have
expectations about Python process output, and we don't want to violate
those expectations unless something is fatally wrong at the "vpython"
layer by default.

BUG=None
TEST=local
  - Reproduced the race condition that was observed.
  - With patch, the race condition could no longer be reproduced.

Review-Url: https://codereview.chromium.org/2845853004